+++ src/Airfield.h
@@ -14,12 +14,12 @@
 
 class Airfield: public Waypoint {
 public:
-	Airfield(const std::string& longName, const std::string& shortName, const std::string& countryCode, const double lat, const double lon, const float alt, const int style, const int rwyDir, const int rwyLen, const double freq, const std::string& descr);
+	Airfield(const std::string& longName, const std::string& shortName, const std::string& countryCode, const double lat, const double lon, const float alt, const int style, const int rwyDir, const int rwyLen, const int freq, const std::string& descr);
 	~Airfield() {}
 
 	inline int GetRunwayDir() const { return runwayDir; }
 	inline int GetRunwayLength() const { return runwayLength; }
-	inline double GetRadioFrequency() const { return radioFreq; }
+	inline int GetRadioFrequency() const { return radioFreq; }
 	inline bool HasRunwayDir() const { return runwayDir > 0; }
 	inline bool HasRunwayLength() const { return runwayLength > 0; }
 	inline bool HasRadioFrequency() const { return radioFreq > 0; }
@@ -27,7 +27,7 @@ public:
 private:
 	int runwayDir;    // [deg]
 	int runwayLength; // [m]
-	double radioFreq; // [MHz]
+	int radioFreq; // [Hz]
 
 };
